+++ Cargo.toml
@@ -19,13 +19,16 @@ serde_json = "1.0.59"
 tokio = { version = "1.14", features = ["full", "time", "rt-multi-thread"] }
 tokio-util = "0.6.9"
 tokio-stream = { version = "0.1.9", features = ["sync"] }
-btleplug = { version = "0.10", features = ["serde"] }
 warp = "0.3"
 nix = "0.23.0"
 regex = "1.5"
 
+[target.'cfg(any(target_os = "macos", target_os = "windows"))'.dependencies]
+btleplug = { version = "0.10", features = ["serde"] }
+
 [target.'cfg(target_os = "linux")'.dependencies]
 rppal = "0.13.1"
+bluer = {version = "0.15.0", features = ["bluetoothd"]}
 
 [build-dependencies]
 protoc-rust = "2.25.2"
